Move the public part of eval.go into eval_api.go
